requirements_lock.txt
absl-py==2.4.0
accelerate==1.13.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
apache-tvm-ffi==0.1.9
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
binpacking==2.0.1
bitsandbytes==0.49.2
blake3==1.0.8
brotli==1.2.0
cachetools==7.0.5
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.13.0
contourpy==1.3.2
cpm-kernels==1.0.11
crcmod==1.7
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.2
cuda-python==12.9.4
cupy-cuda12x==14.0.1
cycler==0.12.1
dacite==1.9.2
datasets==3.6.0
depyf==0.20.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
ffmpy==1.0.0
filelock==3.25.2
flashinfer-python==0.6.4
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.3.0
gguf==0.18.0
googleapis-common-protos==1.73.0
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
grpcio==1.78.0
grpcio-reflection==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
importlib_metadata==8.7.1
interegular==0.3.3
Jinja2==3.1.6
jiter==0.13.0
jmespath==0.10.0
joblib==1.5.3
json_repair==0.58.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaldi-native-fbank==1.22.3
kiwisolver==1.5.0
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.10.0
model-hosting-container-standards==0.1.13
modelscope==1.35.0
mpmath==1.3.0
ms_swift==4.0.2
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
nltk==3.9.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.4.1
nvidia-cutlass-dsl-libs-base==4.4.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.24.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.15
orjson==3.11.7
oss2==2.19.1
outlines_core==0.2.11
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pillow==11.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
quack-kernels==0.3.2
ray==2.54.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rouge==1.0.1
rpds-py==0.30.0
ruff==0.15.6
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.54.0
setproctitle==1.3.7
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.3.2
starlette==0.52.1
supervisor==4.3.0
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.10.0
torch_c_dlpack_ext==0.1.5
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
transformers-stream-generator==0.0.5
triton==3.6.0
trl==0.28.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vllm==0.17.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.6
xgrammar==0.1.29
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
